fix: remove wrong parameter nv
IdealDecompositionPackage and QuasiAlgebraicSet2 used to require two
parameters where the first parameter was a list of variables and the
second parameter was supposed to be the length of this list.

This additional parameter for the list length was seemingly necessary
because of a weakness of the compiler. It no longer is.

In fact, that parameter was wrong conceptually and compilation should
have been rejected by the SPAD compiler. Formerly we had

F     ==> Fraction Integer
Var   ==> OrderedVariableList vl
Expon ==> DirectProduct(nv,NNI)
Dpoly ==> DistributedMultivariatePolynomial(vl,F)
QALG  ==> QuasiAlgebraicSet(F, Var, Expon, Dpoly)

The 4 parameter of QuasiAlgebraicSet is required to be of type
PolynomialCategory(F,Expon,Var). However, the type of
DistributedMultivariatePolynomial(vl,F) is

  PolynomialCategory(F,DirectProduct(#vl,NonNegativeInteger),
                       OrderedVariableList(vl)) with
      reorder: (%,List Integer) -> %

This was a mismatch in the third parameter, namely the provided
DirectProduct(nv,NNI) vs. the expected DirectProduct(#vl,NonNegativeInteger).

git-svn-id: https://fricas.svn.sourceforge.net/svnroot/fricas/trunk@1328 b0c55286-4f34-0410-a049-a1e7e93b0762