de-macroify fricas-lisp.lisp
This commit removes lots of DEFMACROs and replaces them with DEFUNs
when possible. In general, this is preferable for two reasons:

(1) The functions are still available for use as high-order functions.

(2) Functions avoid multiple evaluation of the arguments.

(3) In many cases, the code is simplified.

(4) The assembly code can be readily inspected for performance
    analysis.

This commit also:

- Re-orders some of the definitions so the inline expansions catch.

- Deletes some old SBCL code for constructing specialized vectors.